simplify mgmt commands
